Test ninja-build performance with empty.exe
In order to measure the cost of having ninja doing all process creation
on a single thread (serialized) I made a custom version of ninja that
modifies Subprocess::Start to always run empty.exe. Testing this when
"building" chrome showed that this NOP build took 140 s. This is not
ideal, and is clearly slower than on Linux, but it is fast enough to
make optimization of CreateProcess not as interesting as I had
suspected.

This was tested on a 24-core Z840 although I suspect that the number
of cores doesn't matter much.